backend/requirements.txt
# Existing ML packages
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1+cpu
torchvision==0.20.1+cpu
pillow==10.4.0
numpy==1.26.4
matplotlib==3.9.0
tqdm==4.66.4
scikit-learn==1.5.1
opencv-python-headless==4.10.0.84

# Web3 and blockchain
web3==6.20.0
eth-account==0.11.0
python-dotenv==1.0.0

# Flask API
flask==3.0.3
flask-cors==4.0.1
werkzeug==3.0.3
